Addressing code review changes.
